pyproject.toml
[project]
name = "agenticairetail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11, <3.13"
dependencies = [
    "chromadb>=1.5.5",
    "groq",
    "langchain-openai>=1.1.12",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langgraph>=1.1.3",
    "mcp>=1.26.0",
    "onnxruntime<1.20",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "langchain-groq>=1.1.2",
    "grandalf>=0.8",
    "pinecone>=8.1.0",
]
